% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{find_python}
\alias{find_python}
\title{Find Python versions and Conda Environments}
\usage{
find_python(python = TRUE, conda = TRUE)
}
\arguments{
\item{python}{If TRUE will list available python versions.
Default: TRUE}

\item{conda}{If TRUE will list available conda environments.
Default: TRUE}
}
\description{
This function lists available python versions and conda
environments.
}
