% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_distributions}
\alias{get_distributions}
\title{Get PDF distribution}
\usage{
get_distributions(matrix_interval, distribution_list, asymp_stab)
}
\arguments{
\item{matrix_interval}{Position in the matrix. Example: c(0, 0)}

\item{distribution_list}{list of scipy distribution}

\item{asymp_stab}{asymptotic stability matrix}
}
\description{
This function retrieves the PDF (Probablity Distribution Function)
object from the scipy method 
<scipy.stats._distn_infrastructure.rv_frozen>.
}
\examples{
ax2 <- get_distributions(matrix_interval = c(0, 0), distribution_list, asymp_stab)
}
