% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{PreprocessMatrix}
\alias{PreprocessMatrix}
\title{Preprocess Matrix}
\usage{
PreprocessMatrix(input_file, output_folder = NULL, prefix = NULL,
  max_bound = 10, zero_perturb = FALSE, threads = 1,
  verbose = FALSE)
}
\arguments{
\item{input_file}{Input comma separated file for the jacobian matrix.}

\item{output_folder}{Output folder. A number of files will be created in the form 
‘output_folder/<prefix>_*.npy’}

\item{prefix}{Prefix of output files, if you so choose.}

\item{max_bound}{some of the matrices are unbounded stable towards one end, this 
is the limit the user imposes. Default: 10}

\item{zero_perturb}{Flag to indicated you want to pertub the zero entries. 
Default: FALSE}

\item{threads}{Number of threads to use. Default: 1}

\item{verbose}{TODO}
}
\description{
This script pre-processes a matrix by figuring out what the intervals of asymptotic
stability are, as well as finding which perturbation values lead to a sign switch.
}
