% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{py_depend}
\alias{py_depend}
\title{Install Python Dependencies}
\usage{
py_depend(condaenv = NULL, virtualenv = NULL)
}
\arguments{
\item{condaenv}{Name of conda environment to install python libraries to.
Default: NULL}

\item{virtualenv}{Name of virtual environment to install python libraries to.
Default: NULL}
}
\description{
This function installs needed python libraries into the specified conda
environment OR virtual environment. Should be the same as the one 
specified in set_python.
Required python libraries: matplotlib, numpy, pandas, pathos,
scipy and sympy
On CentOS 7 pandas & scipy may need to be installed with pip install 
from the command line. Will get the error: /lib/libstdc++.so.6: version
`CXXABI_1.3.9' not found
See vingette for more information.
}
\examples{
\dontrun{
# Cond env
py_depend(condaenv = "r-reticulate", 
    virtualenv = NULL)
# virtualenv:
py_depend(virtualenv = "r-reticulate", 
    condaenv = NULL)
}
}
