% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{set_python}
\alias{set_python}
\title{Set Up Python Configuration}
\usage{
set_python(condaenv, version = NULL, verbose = TRUE)
}
\arguments{
\item{condaenv}{Specify conda environment name}

\item{version}{Set path to specific version of python.}

\item{verbose}{TRUE or FALSE. When TRUE, shows python and conda configuration.
Default: TRUE}
}
\description{
This function sets your python version and conda environment.
Run this command before PreprocessMatrix. Install python dependencies in the
same conda environment that you set here.
If conda environment does not exist, this function will make a new one.
}
\examples{
set_python(version = "~/anaconda3/bin/python", condaenv = "r-reticulate", verbose = TRUE)
}
