% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{set_python_virtual}
\alias{set_python_virtual}
\title{Set Up Virtual Environment for Python Configuration}
\usage{
set_python_virtual(virtualenv, version = NULL, verbose = TRUE)
}
\arguments{
\item{virtualenv}{Specify conda environment name}

\item{version}{Set path to specific version of python.}

\item{verbose}{TRUE or FALSE. When TRUE, shows python and conda configuration.
Default: TRUE}
}
\description{
This function sets your python version and Virtual environment.
Run this command before PreprocessMatrix. Install python dependencies in the
same Virtual environment that you set here.
If the virtual environment does not exist, this function will make a new one.
If the virtual environment already exists, no need to set the python version.
When making a new virtual environment, if the python version isn't set, then
your default one will be used.
}
\examples{
set_python_virtual(version = "/usr/bin/python3", virtualenv = "r-reticulate", verbose = TRUE)
}
