% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{GenerateEntryWiseFigures}
\alias{GenerateEntryWiseFigures}
\title{Generate Entry Wise Figures}
\usage{
GenerateEntryWiseFigures(input_folder, prefix = NULL,
  all_numswitch_plots = FALSE, list_of_numswitch_to_plot)
}
\arguments{
\item{input_folder}{Input folder. The location of the files created by PreprocessMatrix.}

\item{prefix}{Prefix of output files, if you so choose.}

\item{all_numswitch_plots}{set to TRUE if you ant to plot all num switch plots 
(potentially very large). Default: FALSE}

\item{list_of_numswitch_to_plot}{List of entries you want visualized with num switch. 
Should be a list of vectors. Example: list(c(0, 0), c(0, 1))}
}
\description{
This function plots the number of mis-predictions versus perturbation
value, overlaid with distribution over stable pertrubation values.
Run after ComputeEntryWisePerturbationExpectation()
}
