% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{get_distributions}
\alias{get_distributions}
\title{Get PDF distribution}
\usage{
get_distributions(matrix_entry, distribution_list, asymp_stab,
  points = 250)
}
\arguments{
\item{matrix_entry}{Position in the matrix. Example: c(0, 0) # matrix entry}

\item{distribution_list}{list of scipy distribution}

\item{asymp_stab}{asymptotic stability interval}

\item{points}{the number of values in x range}
}
\description{
This function retrieves the PDF (Probablity Distribution Function)
object from the scipy method 
<scipy.stats._distn_infrastructure.rv_frozen>.
}
\examples{
ax2 <- get_distributions(matrix_entry = c(0, 0), distribution_list, asymp_stab)
}
