% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{py_depend}
\alias{py_depend}
\title{Install Python Dependencies}
\usage{
py_depend(condaenv)
}
\arguments{
\item{condaenv}{Name of conda environment to install python libraries to.}
}
\description{
This function installs needed python libraries into the specified conda
environment. Should be the same as the one specified in set_python.
Required python libraries: matplotlib, numpy, pandas, pathos,
scipy and sympy
}
\details{
On CentOS 7 pandas & scipy may need to be installed with pip install 
from the command line. Will get the error: /lib/libstdc++.so.6: version
`CXXABI_1.3.9' not found
See vingette for more information.
}
\examples{
py_depend(condaenv = "r-reticulate")
}
